update with cpu-id

+++ nnue/misc.h
@@ -92,7 +92,7 @@ INLINE uint32_t readu_le_u32(const void* p)
 INLINE uint16_t readu_le_u16(const void* p)
 {
 	const auto* q = static_cast<const uint8_t*>(p);
-	return q[0] | (q[1] << 8);
+	return static_cast<uint16_t>(q[0] | (q[1] << 8));
 }
 
 template<typename T1, typename T2, typename T3>
